requirements.txt
pydantic>=2.0
pydantic-ai
python-docx
PyMuPDF
google-generativeai
redis
python-dotenv
streamlit
nest_asyncio